requirements.txt
asgiref==3.8.1
Django==5.2.8
django-bootstrap5==25.1
django-filter==25.2
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.16.1
packaging==25.0
pillow==11.1.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
